feat(ui | sdk): 在制作新声音的界面提供预览声音的按钮, 从后端提供预览播放声音的接口, 供前端实现预览声音的功能。
commit #31

+++ sdk/keySound/keySound.go
@@ -49,8 +49,8 @@ type AudioFilePath struct {
 }
 
 type Cut struct {
-	StartMS int
-	EndMS   int // 当 EndMS 小于或等于 StartMS  时, 不会播放任何声音
+	StartMS int64
+	EndMS   int64 // 当 EndMS 小于或等于 StartMS  时, 不会播放任何声音
 }
 
 // 键音播放器
